move code into classifier agent
